pyproject.toml
[project]
name = "autoocr"
version = "0.1.0"
description = "Modular OCR preprocessing pipeline with conditional image enhancement"
authors = [ { name = "AutoOCR" } ]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
dependencies = [
    "fastapi==0.110.2",
    "uvicorn[standard]==0.29.0",
    "opencv-python==4.9.0.80",
    "numpy==1.26.4",
    "pytesseract==0.3.10",
    "pdf2image==1.17.0",
    "Pillow>=10.3.0",
    "scikit-image==0.22.0",
    "fasttext-wheel==0.9.2",
    "rapidfuzz==3.6.1",
    "python-multipart==0.0.9",
    "reportlab==4.0.9",
    "tqdm==4.66.2"
    ,"PyYAML==6.0.1"
    ,"httpx==0.27.0"
]

[project.scripts]
autoocr = "autoocr.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
